+++ src/Input.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("./Option.mjs").Option} Option */
+
+/**
+ * @typedef {{"auto-focus"?: boolean, disabled?: boolean, max?: string, "max-length"?: number, min?: string, "min-length"?: number, multiple?: boolean, name: string, options?: Option[], pattern?: string, placeholder?: string, "read-only"?: boolean, required?: boolean, step?: string, title?: string, type?: string, value?: string}} Input
+ */