+++ src/Option.mjs
@@ -1,0 +1,3 @@
+/**
+ * @typedef {{disabled?: boolean, label: string, selected?: boolean, title?: string, value: string}} Option
+ */